Break HiveConnector.cpp into multiple files
The HiveConnector.cpp and HiveConnector.h are already very big and
contain many classes. This commit breaks it down into several files.
The HiveDataSource is in analogous to the existing HiveDataSink files.
This is to prepare for the upcoming Iceberg development, such that
the files are easier to maintain.
