+++ velox/connectors/hive/HiveConnector.h
@@ -15,262 +15,10 @@
  */
 #pragma once
 
-#include "velox/connectors/hive/FileHandle.h"
-#include "velox/connectors/hive/HiveConnectorSplit.h"
 #include "velox/connectors/hive/HiveDataSink.h"
-#include "velox/dwio/common/CachedBufferedInput.h"
-#include "velox/dwio/common/IoStatistics.h"
-#include "velox/dwio/common/Reader.h"
-#include "velox/dwio/common/ScanSpec.h"
-#include "velox/dwio/dwrf/writer/Writer.h"
-#include "velox/exec/OperatorUtils.h"
-#include "velox/expression/Expr.h"
-#include "velox/type/Filter.h"
-#include "velox/type/Subfield.h"
+#include "velox/connectors/hive/HiveDataSource.h"
 
 namespace facebook::velox::connector::hive {
-
-class HiveColumnHandle : public ColumnHandle {
- public:
-  enum class ColumnType { kPartitionKey, kRegular, kSynthesized };
-
-  HiveColumnHandle(
-      const std::string& name,
-      ColumnType columnType,
-      TypePtr dataType,
-      std::vector<common::Subfield> requiredSubfields = {})
-      : name_(name),
-        columnType_(columnType),
-        dataType_(std::move(dataType)),
-        requiredSubfields_(std::move(requiredSubfields)) {}
-
-  const std::string& name() const {
-    return name_;
-  }
-
-  ColumnType columnType() const {
-    return columnType_;
-  }
-
-  const TypePtr& dataType() const {
-    return dataType_;
-  }
-
-  // Applies to columns of complex types: arrays, maps and structs.  When a
-  // query uses only some of the subfields, the engine provides the complete
-  // list of required subfields and the connector is free to prune the rest.
-  //
-  // Examples:
-  //  - SELECT a[1], b['x'], x.y FROM t
-  //  - SELECT a FROM t WHERE b['y'] > 10
-  //
-  // Pruning a struct means populating some of the members with null values.
-  //
-  // Pruning a map means dropping keys not listed in the required subfields.
-  //
-  // Pruning arrays means dropping values with indices larger than maximum
-  // required index.
-  const std::vector<common::Subfield>& requiredSubfields() const {
-    return requiredSubfields_;
-  }
-
-  bool isPartitionKey() const {
-    return columnType_ == ColumnType::kPartitionKey;
-  }
-
-  std::string toString() const;
-
-  folly::dynamic serialize() const override;
-
-  static ColumnHandlePtr create(const folly::dynamic& obj);
-
-  static std::string columnTypeName(HiveColumnHandle::ColumnType columnType);
-
-  static HiveColumnHandle::ColumnType columnTypeFromName(
-      const std::string& name);
-
-  static void registerSerDe();
-
- private:
-  const std::string name_;
-  const ColumnType columnType_;
-  const TypePtr dataType_;
-  const std::vector<common::Subfield> requiredSubfields_;
-};
-
-using SubfieldFilters =
-    std::unordered_map<common::Subfield, std::unique_ptr<common::Filter>>;
-
-class HiveTableHandle : public ConnectorTableHandle {
- public:
-  HiveTableHandle(
-      std::string connectorId,
-      const std::string& tableName,
-      bool filterPushdownEnabled,
-      SubfieldFilters subfieldFilters,
-      const core::TypedExprPtr& remainingFilter);
-
-  ~HiveTableHandle() override;
-
-  bool isFilterPushdownEnabled() const {
-    return filterPushdownEnabled_;
-  }
-
-  const SubfieldFilters& subfieldFilters() const {
-    return subfieldFilters_;
-  }
-
-  const core::TypedExprPtr& remainingFilter() const {
-    return remainingFilter_;
-  }
-
-  std::string toString() const override;
-
-  folly::dynamic serialize() const override;
-
-  static ConnectorTableHandlePtr create(
-      const folly::dynamic& obj,
-      void* context);
-
-  static void registerSerDe();
-
- private:
-  const std::string tableName_;
-  const bool filterPushdownEnabled_;
-  const SubfieldFilters subfieldFilters_;
-  const core::TypedExprPtr remainingFilter_;
-};
-
-class HiveConnector;
-
-class HiveDataSource : public DataSource {
- public:
-  HiveDataSource(
-      const RowTypePtr& outputType,
-      const std::shared_ptr<connector::ConnectorTableHandle>& tableHandle,
-      const std::unordered_map<
-          std::string,
-          std::shared_ptr<connector::ColumnHandle>>& columnHandles,
-      FileHandleFactory* fileHandleFactory,
-      velox::memory::MemoryPool* pool,
-      core::ExpressionEvaluator* expressionEvaluator,
-      memory::MemoryAllocator* allocator,
-      const std::string& scanId,
-      folly::Executor* executor);
-
-  void addSplit(std::shared_ptr<ConnectorSplit> split) override;
-
-  void addDynamicFilter(
-      column_index_t outputChannel,
-      const std::shared_ptr<common::Filter>& filter) override;
-
-  std::optional<RowVectorPtr> next(uint64_t size, velox::ContinueFuture& future)
-      override;
-
-  uint64_t getCompletedRows() override {
-    return completedRows_;
-  }
-
-  uint64_t getCompletedBytes() override {
-    return ioStats_->rawBytesRead();
-  }
-
-  std::unordered_map<std::string, RuntimeCounter> runtimeStats() override;
-
-  bool allPrefetchIssued() const override {
-    return rowReader_ && rowReader_->allPrefetchIssued();
-  }
-
-  void setFromDataSource(std::unique_ptr<DataSource> sourceUnique) override;
-
-  int64_t estimatedRowSize() override;
-
-  // Internal API, made public to be accessible in unit tests.  Do not use in
-  // other places.
-  static std::shared_ptr<common::ScanSpec> makeScanSpec(
-      const SubfieldFilters& filters,
-      const RowTypePtr& rowType,
-      const std::vector<const HiveColumnHandle*>& columnHandles,
-      const std::vector<common::Subfield>& remainingFilterInputs,
-      memory::MemoryPool* pool);
-
- protected:
-  virtual uint64_t readNext(uint64_t size) {
-    return rowReader_->next(size, output_);
-  }
-
-  std::unique_ptr<dwio::common::BufferedInput> createBufferedInput(
-      const FileHandle&,
-      const dwio::common::ReaderOptions&);
-
-  virtual std::unique_ptr<dwio::common::RowReader> createRowReader(
-      dwio::common::RowReaderOptions& options) {
-    return reader_->createRowReader(options);
-  }
-
-  std::shared_ptr<HiveConnectorSplit> split_;
-  FileHandleFactory* fileHandleFactory_;
-  dwio::common::ReaderOptions readerOpts_;
-  memory::MemoryPool* pool_;
-  VectorPtr output_;
-  RowTypePtr readerOutputType_;
-  std::unique_ptr<dwio::common::RowReader> rowReader_;
-
- private:
-  // Evaluates remainingFilter_ on the specified vector. Returns number of rows
-  // passed. Populates filterEvalCtx_.selectedIndices and selectedBits if only
-  // some rows passed the filter. If none or all rows passed
-  // filterEvalCtx_.selectedIndices and selectedBits are not updated.
-  vector_size_t evaluateRemainingFilter(RowVectorPtr& rowVector);
-
-  void setConstantValue(
-      common::ScanSpec* FOLLY_NONNULL spec,
-      const TypePtr& type,
-      const velox::variant& value) const;
-
-  void setNullConstantValue(
-      common::ScanSpec* FOLLY_NONNULL spec,
-      const TypePtr& type) const;
-
-  void setPartitionValue(
-      common::ScanSpec* FOLLY_NONNULL spec,
-      const std::string& partitionKey,
-      const std::optional<std::string>& value) const;
-
-  // Clear split_ after split has been fully processed.  Keep readers around to
-  // hold adaptation.
-  void resetSplit();
-
-  void configureRowReaderOptions(dwio::common::RowReaderOptions&) const;
-
-  const RowTypePtr outputType_;
-  // Column handles for the partition key columns keyed on partition key column
-  // name.
-  std::unordered_map<std::string, std::shared_ptr<HiveColumnHandle>>
-      partitionKeys_;
-  std::shared_ptr<dwio::common::IoStatistics> ioStats_;
-  std::shared_ptr<common::ScanSpec> scanSpec_;
-  std::shared_ptr<common::MetadataFilter> metadataFilter_;
-  dwio::common::RowReaderOptions rowReaderOpts_;
-  std::unique_ptr<dwio::common::Reader> reader_;
-  std::unique_ptr<exec::ExprSet> remainingFilterExprSet_;
-  bool emptySplit_;
-
-  dwio::common::RuntimeStatistics runtimeStats_;
-
-  std::shared_ptr<FileHandle> fileHandle_;
-  core::ExpressionEvaluator* expressionEvaluator_;
-  uint64_t completedRows_ = 0;
-
-  // Reusable memory for remaining filter evaluation.
-  VectorPtr filterResult_;
-  SelectivityVector filterRows_;
-  exec::FilterEvalCtx filterEvalCtx_;
-
-  memory::MemoryAllocator* const allocator_;
-  const std::string& scanId_;
-  folly::Executor* executor_;
-};
 
 class HiveConnector : public Connector {
  public:
