+++ velox/exec/tests/utils/HiveConnectorTestBase.cpp
@@ -15,13 +15,10 @@
  */
 
 #include "velox/exec/tests/utils/HiveConnectorTestBase.h"
-#include "velox/common/base/Fs.h"
+
 #include "velox/common/file/FileSystems.h"
-#include "velox/connectors/hive/HiveConnector.h"
 #include "velox/dwio/common/tests/utils/BatchMaker.h"
 #include "velox/dwio/dwrf/reader/DwrfReader.h"
-#include "velox/dwio/dwrf/writer/Writer.h"
-#include "velox/exec/tests/utils/QueryAssertions.h"
 
 namespace facebook::velox::exec::test {
