+++ velox/exec/tests/utils/HiveConnectorTestBase.h
@@ -14,7 +14,9 @@
  * limitations under the License.
  */
 #pragma once
+
 #include "velox/connectors/hive/HiveConnector.h"
+#include "velox/connectors/hive/HiveConnectorSplit.h"
 #include "velox/exec/Operator.h"
 #include "velox/exec/tests/utils/OperatorTestBase.h"
 #include "velox/exec/tests/utils/TempFilePath.h"
